backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
requests==2.32.3
sqlalchemy==2.0.32
psycopg[binary]==3.2.4
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
pyjwt[crypto]==2.9.0
httpx==0.27.2
pydantic-settings==2.6.1
